Win: Fix bezel config mode
Overlapping widgets is underfined and while this worked on Linux, on Windows the bezel edit mode was completely broken, as edit buttons could not be drawn over the bitmap widgets. This is fixed by modifying the Position mode which draws the monitor previews through PaintDC, to suit the needs of the bezel config.